This commit was generated by cvs2svn to compensate for changes in r15283, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@15284 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/debug.c
@@ -290,7 +290,7 @@ void debug_nmi(interrupt_cpu_status_t *cs, CLOCK iclk)
 
 void debug_dma(const char *txt, CLOCK dclk, int num)
 {
-    log_debug("*** DMA %s %10i  %02i", txt, dclk, num);
+    log_debug("*** DMA %s %10ld  %02i", txt, (long)dclk, num);
 }
 
 /*------------------------------------------------------------------------*/
